Move SootNode and its comparator to models module
