+++ build.gradle
@@ -99,6 +99,7 @@ repositories {
 dependencies {
     compile project(":java-maven-project-compiler")
     compile project(":jimple-evosuite-test-generator")
+    compile project(":jimple-library-usage-graph")
     compile project(":jimple-library-usage-graph-generator")
     compile project(":jimple-pattern-filter")
     compile project(":prefix-span-pattern-detector")
