+++ modules/models/jimple-library-usage-graph/gradle.properties
@@ -1,0 +1,2 @@
+mockitoKotlinVersion = 1.5.0
+sootVersion          = 3.1.0
