+++ modules/models/jimple-library-usage-graph/src/main/kotlin/org/cafejojo/schaapi/models/libraryusagegraph/jimple/compare/GeneralizedSootComparator.kt
@@ -1,7 +1,8 @@
-package org.cafejojo.schaapi.usagegraphgenerator.compare
+package org.cafejojo.schaapi.models.libraryusagegraph.jimple.compare
 
+import org.cafejojo.schaapi.common.GeneralizedNodeComparator
 import org.cafejojo.schaapi.common.Node
-import org.cafejojo.schaapi.usagegraphgenerator.SootNode
+import org.cafejojo.schaapi.models.libraryusagegraph.jimple.SootNode
 import soot.Scene
 import soot.Type
 import soot.Unit
