+++ modules/models/jimple-library-usage-graph/src/test/kotlin/org/cafejojo/schaapi/models/libraryusagegraph/jimple/SootNodeTest.kt
@@ -1,4 +1,4 @@
-package org.cafejojo.schaapi.usagegraphgenerator
+package org.cafejojo.schaapi.models.libraryusagegraph.jimple
 
 import com.nhaarman.mockito_kotlin.doReturn
 import com.nhaarman.mockito_kotlin.mock
