+++ modules/pipeline/jimple-evosuite-test-generator/src/test/kotlin/org/cafejojo/schaapi/testgenerator/SootClassGeneratorTest.kt
@@ -2,7 +2,7 @@ package org.cafejojo.schaapi.testgenerator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.catchThrowable
-import org.cafejojo.schaapi.usagegraphgenerator.SootNode
+import org.cafejojo.schaapi.models.libraryusagegraph.jimple.SootNode
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
