+++ modules/pipeline/jimple-library-usage-graph-generator/src/main/kotlin/org/cafejojo/schaapi/usagegraphgenerator/ControlFlowGraphCreator.kt
@@ -1,6 +1,7 @@
 package org.cafejojo.schaapi.usagegraphgenerator
 
 import org.cafejojo.schaapi.common.Node
+import org.cafejojo.schaapi.models.libraryusagegraph.jimple.SootNode
 import soot.Body
 import soot.Unit
 import soot.toolkits.graph.BriefUnitGraph
