+++ modules/pipeline/jimple-library-usage-graph-generator/src/test/kotlin/org/cafejojo/schaapi/usagegraphgenerator/IntegrationTest.kt
@@ -3,6 +3,7 @@ package org.cafejojo.schaapi.usagegraphgenerator
 import com.nhaarman.mockito_kotlin.mock
 import org.assertj.core.api.Assertions.assertThat
 import org.cafejojo.schaapi.common.Node
+import org.cafejojo.schaapi.models.libraryusagegraph.jimple.SootNode
 import org.jetbrains.spek.api.Spek
 import org.jetbrains.spek.api.dsl.describe
 import org.jetbrains.spek.api.dsl.it
