+++ modules/pipeline/jimple-pattern-filter/build.gradle
@@ -4,7 +4,8 @@ repositories {
 
 dependencies {
     compile project(":common")
-    compile project(":jimple-library-usage-graph-generator")
+    compile project(":jimple-library-usage-graph")
+
     compile group: "ca.mcgill.sable", name: "soot", version: sootVersion
 
     testCompile group: "com.nhaarman", name: "mockito-kotlin", version: mockitoKotlinVersion
