+++ modules/pipeline/jimple-pattern-filter/src/main/kotlin/org/cafejojo/schaapi/patternfilter/IncompleteInitPatternFilter.kt
@@ -2,7 +2,7 @@ package org.cafejojo.schaapi.patternfilter
 
 import org.cafejojo.schaapi.common.Node
 import org.cafejojo.schaapi.common.PatternFilter
-import org.cafejojo.schaapi.usagegraphgenerator.SootNode
+import org.cafejojo.schaapi.models.libraryusagegraph.jimple.SootNode
 import soot.jimple.InvokeStmt
 import soot.jimple.internal.JSpecialInvokeExpr
